refactor: enhance spotify optimistic UI architecture
- Moved duplicated context logic out of AudioContext to hooks/useAudioPreference.ts
- Refactored SpotifyControls to fully utilize useSpotifyVolume, removing duplicated local state
- Extracted scaling logic in VolumeSlider to leverage MUI theme.spacing instead of a manual scaleFactor
- Removed stale constants and updated syncing locks
- Updated spotifyPlayerManager executeOptimistic to emit failure events, unlocking useSpotifyVolume UI freeze early
- Cleaned up redundant verbose comments in SpotifyControls
