+++ app/client/control/components/SpotifyControls.tsx
@@ -11,7 +11,6 @@ import Select from '@mui/material/Select'
 import Typography from '@mui/material/Typography'
 import { useRouter } from 'next/navigation'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import throttle from 'lodash.throttle'
 import { clampVolume } from '@/utils/audioManager'
 import { useWebSocket } from '@/context/WebSocketContext'
 import { useSpotifyCommand } from '@/hooks/useSpotifyCommand'
@@ -33,10 +32,6 @@ const SpotifyControls = () => {
   const { devices = [] } = spotifyData
   const lastSentVolumeRef = useRef<string | null>(null)
   const [selectedDeviceId, setSelectedDeviceId] = useState<string>('')
-  const [isSliding, setIsSliding] = useState(false)
-  const prevActiveIdRef = useRef<string | undefined>(undefined)
-  const lastVolumeSyncTimeRef = useRef<number>(0)
-  const hasPendingSendRef = useRef<boolean>(false)
   const [optimisticIsPlaying, setOptimisticIsPlaying] = useState<
     boolean | null
   >(null)
@@ -79,27 +74,6 @@ const SpotifyControls = () => {
     [connectionStatus, resolveTargetDeviceId, executeSpotify]
   )
 
-  const throttledSendVolume = useMemo(
-    () =>
-      throttle((val: number) => {
-        sendVolumeCommand(val)
-      }, 200),
-    [sendVolumeCommand]
-  )
-
-  useEffect(() => {
-    return () => {
-      throttledSendVolume.cancel()
-    }
-  }, [throttledSendVolume])
-
-  const handleThrottledVolumeChange = useCallback(
-    (val: number) => {
-      throttledSendVolume(val)
-    },
-    [throttledSendVolume]
-  )
-
   const {
     displayVolume,
     isMuted,
@@ -110,14 +84,6 @@ const SpotifyControls = () => {
     spotifyData.playback.volume_percent,
     spotifyData.playback.isMuted,
     sendVolumeCommand
-  )
-
-  const handleVolumeSlide = useCallback(
-    (val: number) => {
-      handleVolumeChange(val)
-      handleThrottledVolumeChange(val)
-    },
-    [handleVolumeChange, handleThrottledVolumeChange]
   )
 
   const handleTrackSelect = (uri: string) => {
@@ -212,14 +178,12 @@ const SpotifyControls = () => {
         command === 'NEXT' ||
         command === 'PREVIOUS'
       ) {
-        // Optimistic UI update for Play/Pause
         if (command === 'PLAY') {
           setOptimisticIsPlaying(true)
         } else if (command === 'PAUSE') {
           setOptimisticIsPlaying(false)
         }
 
-        // Clear existing timer if any
         if (playbackGraceTimerRef.current) {
           clearTimeout(playbackGraceTimerRef.current)
         }
@@ -235,7 +199,6 @@ const SpotifyControls = () => {
     [sendSpotifyCommand]
   )
 
-  // Cleanup timers on unmount
   useEffect(() => {
     return () => {
       if (playbackGraceTimerRef.current) {
@@ -243,55 +206,6 @@ const SpotifyControls = () => {
       }
     }
   }, [])
-
-  const handleVolumeChange = useCallback(
-    (val: number) => {
-      setIsSliding(true)
-      setVolume(val)
-      if (connectionStatus !== 'Connected') {
-        const now = Date.now()
-        // Throttle warning to once every 3 seconds to avoid spam during sliding
-        if (now - lastWarningTimeRef.current > 3000) {
-          showWarning('Changes not saved: Offline')
-          lastWarningTimeRef.current = now
-        }
-      }
-    },
-    [connectionStatus, showWarning, setVolume]
-  )
-
-  const sendVolumeCommand = useCallback(
-    (value: number) => {
-      if (connectionStatus !== 'Connected') return
-      const targetDeviceId = resolveTargetDeviceId()
-
-      // Prevent sending volume command if no device is targeted
-      if (!targetDeviceId) return
-
-      const sanitized = clampVolume(value)
-      const messageKey = `${targetDeviceId}:${sanitized}`
-      if (lastSentVolumeRef.current === messageKey) return
-
-      hasPendingSendRef.current = true
-      lastVolumeSyncTimeRef.current = Date.now()
-
-      executeSpotify('SET_VOLUME', {
-        volume: sanitized,
-        deviceId: targetDeviceId,
-      })
-
-      lastSentVolumeRef.current = messageKey
-    },
-    [connectionStatus, resolveTargetDeviceId, executeSpotify]
-  )
-
-  const handleVolumeChangeCommitted = useCallback(
-    (val: number) => {
-      setIsSliding(false)
-      sendVolumeCommand(val)
-    },
-    [sendVolumeCommand]
-  )
 
   useEffect(() => {
     if (connectionStatus !== 'Connected') {
@@ -367,7 +281,7 @@ const SpotifyControls = () => {
             <VolumeSlider
               volume={displayVolume}
               muted={isMuted}
-              onVolumeChange={handleVolumeSlide}
+              onVolumeChange={handleVolumeChange}
               onVolumeChangeCommitted={handleVolumeChangeCommitted}
               onToggleMute={handleToggleMute}
               showValue
