+++ constants/spotify.ts
@@ -8,6 +8,5 @@ export const HRM_WEB_PLAYER_NAME = 'HRM Web Player'
 export const SPOTIFY_DEFAULT_TOKEN_EXPIRY_S = 3600
 
 // Centralized constants for Spotify integration
-export const VOLUME_SYNC_GRACE_PERIOD_MS = 3000
 export const SPOTIFY_BRAND_COLOR = '#1DB954'
 export const SYNC_LOCK_DURATION = 2000
